Made more uniform with entire document
All references to other elements have the word element after the "elements' title, without  < > around them, when it is mentioned/refrenced in a tag, attribute, elements definition. Adding the word just makes it more uniform, since comma is optional took it away as it looked crowded with the extra word, can be added back another option may be to not use the word element after 'picture' or 'audio' as it could be assumed placed after 'video', or state they are elements before the listing, but that wouldn't be uniform, nor consitant with other definitions.